Create new module for CLICommand
The hope is that commands over time and any related logic can
be moved over to this module as uncoupling happens. Currently
the ServiceCommand, etc. are too coupled to extract these out
yet.
